Separate DFP Data Lifecycle from Caching
